Add descriptor.json and adjust package structure
A new file, descriptor.json, has been added to the react-remote-intrig package. This defines the structure and code generation of the package. Also, the structure of the 'Sample' in Intrig package has been adjusted by renaming 'getPet.ts' to 'getPet.ts.txt' and url attribute was added to the useCallIntegration function in getPet.ts.txt. This is to provide clarity on transformation steps.
Tweaked package.json and tsconfig.base.json to include proper references to the Intrig package. This was necessary for correct builds and enable usage of the templates defined in descriptor.json. Deletion of a file has been inferred from a diff but without a filename, the specific file deleted can't be mentioned.