Remove unnecessary files and configurations
The commit deletes a number of unnecessary files and configurations from NX projects in the repository:
intrig-cli-e2e, intrig-common-e2e, and intrig-openapi3-e2e.
Code that was not used anywhere (resolving definitions, generating interfaces, utility functions, and some tests) was also deleted.
This cleanup is a part of a major refactor with a goal of simplifying the codebase and making it easier to maintain and extend in future development.