Add file and import generation in interface and hook templates
Updated the interfaceTemplate.ts.hbs and hookTemplate.ts.hbs files to generate the necessary files and imports accordingly. This update provides a more streamline way of creating components and handling imports on react-rest-ts-intrig library. The changes ensure that all necessary components are generated when instances are created, providing better management on the generation of unique imports and types.